modbus功能添加，待测试

+++ app/app_buf.h
@@ -8,13 +8,13 @@
 
 typedef void *app_buf_handle;
 
-app_buf_handle app_buf_init(uint8_t capacity);
+app_buf_handle app_buf_init(int capacity);
 
 void app_buf_deinit(app_buf_handle buf);
 
 com_status_e app_buf_write(app_buf_handle buf, char *data, uint8_t len);
 
-com_status_e app_buf_read(app_buf_handle buf, char *pdata, uint8_t data_capacity, uint8_t *real_len);
+com_status_e app_buf_read(app_buf_handle buf, char *pdata, int data_capacity, uint8_t *real_len);
